# GeneratedRESTApiClient/api/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- description: Generated server url
  url: http://localhost:8090
paths:
...
  /example/post-test:
    post:
      operationId: postTest
      requestBody:
        content:
          application/json:
            schema:
              type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: string
          description: OK
      tags:
      - rest-api-example
      x-contentType: application/json
      x-accepts: '*/*'
...